Check that the workspace isn't available/accessible on the backend
